+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 ## [Unreleased]
+### Fixed
+- Fixed ditto marks (") appearing in the ae7q call command.
 
 
 ## [v2.1.0] - 2020-01-04