+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Changed
 - Changelog command to accept a version as argument.
+- The qrz command can now link to a QRZ page instead of embedding the data with the `--link` flag.
 
 
 ## [2.1.0] - 2020-01-04