+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 ## [Unreleased]
+### Added
+- Added Trustee field to qrz command for club callsigns.
 ### Changed
 - Changelog command to accept a version as argument.
 - The qrz command can now link to a QRZ page instead of embedding the data with the `--link` flag.