+++ CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Changelog command to accept a version as argument.
 - The qrz command can now link to a QRZ page instead of embedding the data with the `--link` flag.
+- All currently-available pools can now be accessed by the `hamstudy` command.
+- The `hamstudy` command now uses the syntax `?hamstudy <country> <pool>`.
+- Replaced `hamstudyanswer` command with answering by reaction.
 ### Fixed
 - Fixed ditto marks (") appearing in the ae7q call command.
 - Fixed issue where incorrect table was parsed in ae7q call command.