+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - All currently-available pools can now be accessed by the `hamstudy` command.
 - The `hamstudy` command now uses the syntax `?hamstudy <country> <pool>`.
 - Replaced `hamstudyanswer` command with answering by reaction.
+- Removed all generic error handling from commands.
 ### Fixed
 - Fixed ditto marks (") appearing in the ae7q call command.
 - Fixed issue where incorrect table was parsed in ae7q call command.